fix : 코드리뷰 반영2 -> 라우터 방식 변경
