+++ week4/forecast/src/pages/Home.jsx
@@ -1,9 +1,14 @@
 import React from "react";
+import { Outlet } from "react-router-dom";
 import PageLayout from "../components/common/PageLayout";
 
 const Home = () => {
   // 기본 홈이 되는 페이지
-  return <PageLayout></PageLayout>;
+  return (
+    <PageLayout>
+      <Outlet />
+    </PageLayout>
+  );
 };
 
 export default Home;
